+++ ABB_BF/ABB_BF.DAL/Entities/Grant.cs
@@ -6,7 +6,7 @@
         public string College { get; set; }
         public string EducationLevel { get; set; }
         public string EducationForm { get; set; }
-        public float AverageMarks { get; set; }
+        public decimal AverageMarks { get; set; }
         public string Speciality { get; set; }
         public string? OtherGrants { get; set; }
     }
